--- a/psc-docs/Main.hs
+++ b/psc-docs/Main.hs
@@ -8,6 +8,7 @@ import Control.Arrow (first, second)
 import Control.Category ((>>>))
 import Control.Monad.Writer
 import Data.Text (Text)
+import qualified Data.Text as T
 import Data.Function (on)
 import Data.List
 import Data.Maybe (fromMaybe)
@@ -82,12 +83,12 @@ docgen (PSCDocsOptions fmt inputGlob output) = do
   where
   guardMissing [] = return ()
   guardMissing [mn] = do
-    hPutStrLn stderr ("psc-docs: error: unknown module \"" ++ P.runModuleName mn ++ "\"")
+    hPutStrLn stderr ("psc-docs: error: unknown module \"" ++ T.unpack (P.runModuleName mn) ++ "\"")
     exitFailure
   guardMissing mns = do
     hPutStrLn stderr "psc-docs: error: unknown modules:"
     forM_ mns $ \mn ->
-      hPutStrLn stderr ("  * " ++ P.runModuleName mn)
+      hPutStrLn stderr ("  * " ++ T.unpack (P.runModuleName mn))
     exitFailure
 
   successOrExit :: Either P.MultipleErrors a -> IO a
@@ -186,11 +187,11 @@ parseItem :: String -> DocgenOutputItem
 parseItem s = case elemIndex ':' s of
   Just i ->
     s # splitAt i
-        >>> first P.moduleNameFromString
+        >>> first (P.moduleNameFromString . T.pack)
         >>> second (drop 1)
         >>> IToFile
   Nothing ->
-    IToStdOut (P.moduleNameFromString s)
+    IToStdOut (P.moduleNameFromString (T.pack s))
 
   where
   infixr 1 #