--- a/purescript.cabal
+++ b/purescript.cabal
@@ -206,6 +206,7 @@ library
                      Language.PureScript.Parser.Kinds
                      Language.PureScript.Parser.State
                      Language.PureScript.Parser.Types
+                     Language.PureScript.Prelude
                      Language.PureScript.Pretty
                      Language.PureScript.Pretty.Common
                      Language.PureScript.Pretty.JS