--- a/purescript.cabal
+++ b/purescript.cabal
@@ -206,7 +206,6 @@ library
                      Language.PureScript.Parser.Kinds
                      Language.PureScript.Parser.State
                      Language.PureScript.Parser.Types
-                     Language.PureScript.Prelude
                      Language.PureScript.Pretty
                      Language.PureScript.Pretty.Common
                      Language.PureScript.Pretty.JS