--- a/src/Language/PureScript/Constants.hs
+++ b/src/Language/PureScript/Constants.hs
@@ -1,7 +1,7 @@
 -- | Various constants which refer to things in the Prelude
 module Language.PureScript.Constants where
 
-import Protolude
+import Language.PureScript.Prelude
 
 import Language.PureScript.Names