--- a/src/Language/PureScript/Names.hs
+++ b/src/Language/PureScript/Names.hs
@@ -5,7 +5,7 @@
 --
 module Language.PureScript.Names where
 
-import Protolude
+import Language.PureScript.Prelude
 
 import Control.Monad.Supply.Class