… examples/passing/StringEscapes.purs
… src/Language/PureScript/CoreFn/Expr.hs
… src/Language/PureScript/CoreFn/ToJSON.hs
… src/Language/PureScript/Docs/AsMarkdown.hs
… src/Language/PureScript/Docs/Convert.hs
… src/Language/PureScript/Docs/Convert/ReExports.hs
… src/Language/PureScript/Docs/Convert/Single.hs
… src/Language/PureScript/Docs/Render.hs
… src/Language/PureScript/Docs/RenderedCode/Render.hs
… src/Language/PureScript/Docs/RenderedCode/Types.hs
… src/Language/PureScript/Docs/Types.hs
… src/Language/PureScript/Errors.hs
… src/Language/PureScript/Errors/JSON.hs
… src/Language/PureScript/Externs.hs
… src/Language/PureScript/Ide.hs
… src/Language/PureScript/Ide/Imports.hs
… src/Language/PureScript/Interactive.hs
… src/Language/PureScript/Interactive/Completion.hs
… src/Language/PureScript/Interactive/Parser.hs
… src/Language/PureScript/Interactive/Printer.hs
… src/Language/PureScript/Linter.hs
… src/Language/PureScript/Linter/Exhaustive.hs
… src/Language/PureScript/Linter/Imports.hs
… src/Language/PureScript/Make.hs
… src/Language/PureScript/Parser/Common.hs
… src/Language/PureScript/Parser/Declarations.hs
… src/Language/PureScript/Parser/Lexer.hs
… src/Language/PureScript/Parser/Types.hs
… src/Language/PureScript/Prelude.hs
… src/Language/PureScript/Pretty/Common.hs