Fixing #547.

I think I overdid it with the changes inside `Make` and there still are some spots with lots of converting.

To get rid of the conversions inside the pretty printer I'm planning on porting a `Text` version of `boxes`, and the conversions inside `Docs` might need some help from @hdgarrood.

My plans for this PR:

1. Release v0.10.2
2. Fix the outstanding issues around `Make`
3. Merge this
4. Try it on lots of existing PS code and see what breaks...

Because this PR touches the entire codebase it will create lots of merge conflicts ff it stays open too long. So I'd suggest we reject it outright or work on getting it merged asap.

The commit history is a bit messy, but I think taking it file-wise is still besser than just squashing all the commits.

EDIT: on top of #2417, so it should be merged after that one