@hdgarrood 2016-11-05
Re boxes, you might find this thread https://github.com/treeowl/boxes/pull/1 interesting.