@hdgarrood 2016-11-05
Actually I think I'll have a stab at generalising `boxes` now, unless you're already looking at it?