@paf31 2016-11-06
Looks great so far! I'm interested to see what sort of speed improvement we can get with a modified version of `boxes` - I suspect it'll be pretty good.

I won't review all of this. I assume it's mostly a mechanical translation? `pack`ing and `unpack`ing small strings here and there shouldn't be a big deal in terms of performance, but we should maybe be careful to avoid transforming back and forth if the strings are potentially large. Obviously, we should just profile and see if there are any hotspots before merging.