@kRITZCREEK 2016-11-06
Because it uses the Plattform specific encoding, which is different between
Windows and *ix.

Am 06.11.2016 8:38 vorm. schrieb "Rémi Vion" <notifications@github.com>:

> *@rvion* commented on this pull request.
> ------------------------------
>
> In src/System/IO/UTF8.hs
> <https://github.com/purescript/purescript/pull/2418#pullrequestreview-7328296>
> :
>
> > @@ -4,11 +4,21 @@ import Prelude.Compat
>
>  import qualified Data.ByteString as BS
>  import qualified Data.ByteString.UTF8 as UTF8
> +import           Data.Text (Text)
> +import qualified Data.Text.Encoding as TE
> +
> +readUTF8FileT :: FilePath -> IO Text
> +readUTF8FileT inFile =
> +  fmap TE.decodeUtf8 (BS.readFile inFile)
>
> why not use https://hackage.haskell.org/package/text-1.2.2.1/docs/
> Data-Text-IO.html instead ?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/purescript/purescript/pull/2418#pullrequestreview-7328296>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AF5xVbJAv3h_u2Z4B5573W-a-V5ZdyLCks5q7YP4gaJpZM4KqXhU>
> .
>