@kRITZCREEK 2016-11-08
The reason for the test failure is because `passing/StringEscapes.purs` is encoded with the windows encoding and not UTF8. It says `dos` in my emacs status line at least. I'm a little lost tbh...