@andyarvanitis 2016-11-14
Please make sure to measure the performance gains on this one, because it looks like it will be painful 😧