@kRITZCREEK 2016-11-14
So I've tracked the test failure down to the "surrogate pairs" test case in `StringEscapes.purs`. Does that ring a bell for someone? 

@andyarvanitis I'll run some benchmarks for this against slamdata soon.