@kRITZCREEK 2016-11-16
slamdata current master:

```
	psc +RTS -N -p -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =      391.24 secs   (391238 ticks @ 1000 us, 1 processor)
	total alloc = 302,928,121,944 bytes  (excludes profiling overheads)
```

slamdata with this PR:

```
	psc +RTS -N -p -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =      325.54 secs   (325535 ticks @ 1000 us, 1 processor)
	total alloc = 286,898,050,984 bytes  (excludes profiling overheads)
```

I've also got positive results for `pscid` and `purescript-halogen` but I'm still working on making these benchmarks easily reproducable.