@kRITZCREEK 2016-11-17
@michaelficarra Could you maybe take a look at this? The test failure is the `surrogatePair` case in `StringEscapes.purs`. `"\x1D306"` seems to work fine, but `"\xD834\xDF06"` ends up as two characters instead of one. I've tried using the decoder we used before this PR and `pack`ing up the resulting string, but to no avail.