@hdgarrood 2016-11-17
Perhaps this is due to the way Text is represented? https://hackage.haskell.org/package/text-1.2.2.1/docs/Data-Text-Internal.html