@hdgarrood 2016-11-17
Yeah, what I think is happening is that Text (as documented above), since it uses UTF-16, cannot represent code points in the range 0xD8000 to 0xDFFFF; if you try to `pack` a `String` that contains those code points, it will convert them to the replacement character, 0xFFFD (or `\65533`, as it appears in the `Show Char` instance). I think the problem is here:

```haskell
  parseStringLiteral :: Lexer u Text
  parseStringLiteral = blockString <|> T.pack <$> PT.stringLiteral tokenParser
    where
    delimiter   = P.try (P.string "\"\"\"")
    blockString = delimiter *> (T.pack <$> P.manyTill P.anyChar delimiter)
```

That is, line 254 of `Language.PureScript.Parser.Lexer`.

I think we have two options:
* Modify this code to make it work as it did before, i.e. before constructing a `Text` value that contains the data that a string literal represents, we attempt to 'interpret' surrogate pairs into real code points. This might involve writing our own version of Parsec's `stringLiteral`.
* Start throwing errors inside the parser if surrogate pairs are encountered in string literals.