@hdgarrood 2016-11-17
AFAICT the easiest approach would simply be to switch out the two uses of `Data.Text.pack` in the above code with a different version of `pack` that preserves valid surrogate pairs, but I've been looking a bit and I have not been able to find such a function  yet.