@kRITZCREEK 2016-11-23
## Benchmarking

I ran benchmarks for master and this PR across `pscid`, `purescript-halogen` and `slamdata`and uploaded my script so it can be used to reproduce my results: https://github.com/kRITZCREEK/ps-benchmarks.

I'll not upload the full `.prof` files as they are about 30mb and just copy the header inside this comment instead. If you want the full files just message me.

I only created heap captures for pscid and halogen, since compiling slamdata with heap capture takes a really long time.
These numbers are from running `psc` on my MBP 2013 with 16GB and I'd appreciate it if some of you ran the benchmarks as well.

### pscid

#### This PR:

```
	Wed Nov 23 11:50 2016 Time and Allocation Profiling Report  (Final)

	   psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =       48.43 secs   (48427 ticks @ 1000 us, 1 processor)
	total alloc = 33,919,690,264 bytes  (excludes profiling overheads)
```
![pscid-new](https://cloud.githubusercontent.com/assets/6189397/20560591/68ffaa6a-b17a-11e6-8437-bf2e20f1ca5a.jpg)

#### Current master:

```
	Wed Nov 23 12:54 2016 Time and Allocation Profiling Report  (Final)

	   psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =       58.06 secs   (58057 ticks @ 1000 us, 1 processor)
	total alloc = 35,607,088,400 bytes  (excludes profiling overheads)
```
![pscid-master](https://cloud.githubusercontent.com/assets/6189397/20561053/c367461e-b17c-11e6-9797-8f4010b0f3c3.jpg)

### purescript-halogen

#### This PR:

```
	Wed Nov 23 11:54 2016 Time and Allocation Profiling Report  (Final)

	   psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =       36.14 secs   (36144 ticks @ 1000 us, 1 processor)
	total alloc = 37,371,593,504 bytes  (excludes profiling overheads)
```

![halogen-new](https://cloud.githubusercontent.com/assets/6189397/20560606/7ae4e952-b17a-11e6-92d7-03b5984d4742.jpg)

#### Current master:

```
	Wed Nov 23 12:58 2016 Time and Allocation Profiling Report  (Final)

	   psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =       38.41 secs   (38415 ticks @ 1000 us, 1 processor)
	total alloc = 39,104,490,608 bytes  (excludes profiling overheads)

```
![halogen-master](https://cloud.githubusercontent.com/assets/6189397/20561152/54e1e3a6-b17d-11e6-9b99-f92d03ee9ee9.jpg)

### slamdata

#### This PR:

```
	Wed Nov 23 12:22 2016 Time and Allocation Profiling Report  (Final)

	   psc +RTS -N -p -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =      388.91 secs   (388909 ticks @ 1000 us, 1 processor)
	total alloc = 352,121,911,672 bytes  (excludes profiling overheads)
```

#### Current master:

```
	Wed Nov 23 13:17 2016 Time and Allocation Profiling Report  (Final)

	   psc +RTS -N -p -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

	total time  =      397.79 secs   (397788 ticks @ 1000 us, 1 processor)
	total alloc = 316,814,098,856 bytes  (excludes profiling overheads)
```