@kRITZCREEK 2016-11-23
Interestingly when running these benchmarks on my Arch Linux virtual machine with 8GB of RAM I end up with results that show a lot more improvement across all projects:

### pscid

#### This PR:

```
Wed Nov 23 20:27 2016 Time and Allocation Profiling Report  (Final)

    psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

total time  =       43.32 secs   (43323 ticks @ 1000 us, 1 processor)
total alloc = 30,561,449,216 bytes  (excludes profiling overheads)
```

#### Current master:

```
Wed Nov 23 22:12 2016 Time and Allocation Profiling Report  (Final)

    psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

total time  =       58.17 secs   (58168 ticks @ 1000 us, 1 processor)
total alloc = 33,457,472,024 bytes  (excludes profiling overheads)
```

### purescript-halogen

#### This PR:

```
Wed Nov 23 20:30 2016 Time and Allocation Profiling Report  (Final)

   psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

total time  =       31.12 secs   (31119 ticks @ 1000 us, 1 processor)
total alloc = 33,083,578,280 bytes  (excludes profiling overheads)
```

#### Current master:

```
Wed Nov 23 22:16 2016 Time and Allocation Profiling Report  (Final)

    psc +RTS -N -p -hc -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

total time  =       34.33 secs   (34330 ticks @ 1000 us, 1 processor)
total alloc = 36,409,976,080 bytes  (excludes profiling overheads)
```

### slamdata

#### This PR:
```
Wed Nov 23 20:49 2016 Time and Allocation Profiling Report  (Final)

   psc +RTS -N -p -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

total time  =      325.20 secs   (325199 ticks @ 1000 us, 1 processor)
total alloc = 287,474,895,088 bytes  (excludes profiling overheads)
```

Current master:
```
Wed Nov 23 22:36 2016 Time and Allocation Profiling Report  (Final)

    psc +RTS -N -p -RTS src/**/*.purs bower_components/purescript-*/src/**/*.purs

total time  =      390.35 secs   (390352 ticks @ 1000 us, 1 processor)
total alloc = 303,540,835,824 bytes  (excludes profiling overheads)

```